Attempted fix for voids at edges of H2O remap area
It was noticed that voids (zero values) were occurring in a ring around
the remapped contrail at each time step. This was because the remap from
the old grid to the new includes a small number of cells where only part
of the destination cell is covered by a source cell. These locations would
become non-zero during remapping (preventing them being overwritten), but
would not include the contributions of non-contrail water. This is now fixed
by tracking the fraction of each output grid cell which was covered by input
grid cells. The unattributed fraction has background water added.