Water vapor now advects, and water diffusion is relative to BC
Horizontal advection (due to wind shear) of the water vapor
field is now enabled. This revealed an issue whereby diffusion
of water vapor in the boundary resulted in a stripe of excessive
water being advected into the domain. It also appeared that the
diffusion operation was not correctly applying the boundary
conditions from the meteorological data, and there remained an
ongoing issue regarding diffusion of the background water field.
Rather than trying to resolve these, the treatment of water for
diffusion has instead been changed. Diffusion of water (only) is
calculated after subtracting the boundary condition, which is then
added back in. This changes the boundary condition to zero, which
appears to be reasonably well handled. So far only a background
which does not vary in time has been tested; it remains to be seen
what happens if the background conditions do vary. However this at
least allows advection of water vapor to be enabled (which it
now is) as well as diffusion.