Fix double counting of ice mass in water vapour when processing EPM results (#59)
When the early plume model output is used to initialize the mature plume model, any ice is mapped on to the grid as aerosol. We also calculate water released from the fuel, but since we do not account for the water already taken up as ice we will end up double-counting. Subtracting the ice mass from the released water vapour resolves the issue. The net effect can be negative (if e.g. the air was initially supersaturated) but this is by design - the total water vapour in the air after this operation should still be non-negative.